# be/src/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Swagger Malaria - OpenAPI 3.0"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/v1"
    }
  ],
...
  "paths": {
...
    "/admin/add-user": {
      "post": {
        "tags": ["admin"],
        "summary": "Add new user to system",
        "description": "Admin system add new user (doctor, medical technoligist, employee,...)",
        "operationId": "addUser",
        "requestBody": {
          "description": "Passing user info",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "chung.nguyen_quang"
          },
          "password": {
            "type": "string",
            "example": "123123123"
          },
          "name": {
            "type": "string",
            "example": "Chung Nguyen"
          },
          "role": {
            "type": "string",
            "description": "Employee/Doctor/Medical Technologist",
            "example": "Doctor",
            "enum": ["Doctor", "Medical Technologist"]
          },
          "avatar": {
            "type": "string",
            "example": "http://example.png"
          },
          "info": {
            "type": "object",
            "example": {
              "age": 22,
              "department": "neurology"
            }
          }
        },
        "xml": {
          "name": "order"
        }
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "name": "Authorization",
        "in": "header",
        "type": "apiKey",
        "description": "Enter the token with the `Bearer: ` prefix, e.g. `Bearer abcde12345`"
      }
    }
  },
  "security": [{ "bearerAuth": [] }]
}